@@ -26,15 +26,13 @@
     "build:typedoc": "./tool/typedoc-build.sh",
     "build:11ty": "NODE_OPTIONS='-r ts-node/register' eleventy",
     "watch:11ty": "npm run build:11ty -- --serve --incremental",
-    "tsc": "tsc",
-    "gts": "gts fix",
-    "gts:ci": "gts lint",
-    "prettier": "prettier --write '**/*.{json,yml,scss}'",
-    "prettier:ci": "prettier --check '**/*.{json,yml,scss}'",
-    "stylelint": "stylelint 'source/assets/sass/*.{css,scss}' --fix",
-    "stylelint:ci": "stylelint 'source/assets/sass/*.{css,scss}'",
-    "lint": "run-s prettier gts stylelint tsc",
-    "lint:ci": "run-s prettier:ci gts:ci stylelint:ci tsc",
+    "check": "run-s check:gts check:tsc check:stylelint",
+    "check:gts": "gts check",
+    "check:stylelint": "stylelint 'source/assets/sass/*.{css,scss}'",
+    "check:tsc": "tsc --noEmit",
+    "fix": "run-s fix:gts fix:stylelint",
+    "fix:gts": "gts fix",
+    "fix:stylelint": "stylelint 'source/assets/sass/*.{css,scss}' --fix",
     "test": "echo \"Error: no test specified\" && exit 1",
     "checklinks:internal": "npx --yes --package=hyperlink -- 'hyperlink --canonicalroot https://sass-lang.com/ --root _site/ -r -p -i --skip \"fragment-redirect\" --skip \"types.Map\" _site/index.html'",
     "checklinks:external": "npx --yes --package=broken-link-checker -- 'blc -ro --exclude=twitter.com --exclude=mvnrepository.com --exclude=drupal.org --exclude=web.archive.org https://sass-lang.com'"
@@ -46,7 +44,15 @@
     "@babel/core": "^7.25.2",
     "@babel/preset-env": "^7.25.4",
     "@babel/preset-typescript": "^7.24.7",
+    "@codemirror/autocomplete": "^6.18.0",
+    "@codemirror/commands": "^6.6.1",
+    "@codemirror/lang-css": "^6.3.0",
     "@codemirror/lang-sass": "^6.0.2",
+    "@codemirror/language": "^6.10.2",
+    "@codemirror/lint": "^6.8.1",
+    "@codemirror/state": "^6.4.1",
+    "@codemirror/view": "^6.33.0",
+    "@lezer/highlight": "^1.2.1",
     "@rollup/plugin-babel": "^6.0.4",
     "@rollup/plugin-commonjs": "^26.0.1",
     "@rollup/plugin-inject": "^5.0.5",
@@ -65,11 +71,12 @@
     "@types/prismjs": "^1.26.4",
     "@types/seedrandom": "^3.0.8",
     "@types/semver": "^7.5.8",
-    "@typescript-eslint/eslint-plugin": "^8.4.0",
+    "@typescript-eslint/eslint-plugin": "^8.5.0",
     "cheerio": "^1.0.0-rc.12",
     "codemirror": "^6.0.1",
     "date-fns": "^3.6.0",
     "deep-equal": "^2.2.3",
+    "eslint-plugin-import": "^2.30.0",
     "eslint-plugin-prettier": "^5.2.1",
     "gts": "^5.3.1",
     "immutable": "^4.3.7",
@@ -82,7 +89,7 @@
     "lorem-ipsum": "^2.0.8",
     "markdown-it": "^14.1.0",
     "markdown-it-anchor": "^9.1.0",
-    "markdown-it-attrs": "^4.1.6",
+    "markdown-it-attrs": "^4.2.0",
     "markdown-it-deflist": "^3.0.0",
     "markdown-it-footnote": "^4.0.0",
     "netlify-plugin-11ty": "^1.4.0",
@@ -96,7 +103,7 @@
     "stylelint-config-standard-scss": "^11.1.0",
     "truncate-html": "^1.1.2",
     "ts-node": "^10.9.2",
-    "typescript": "^5.5.4",
+    "typescript": "^5.6.2",
     "typogr": "^0.6.8"
   },
   "dependencies": {
