@@ -1,3 +1,4 @@
+/** The 11ty page object. */
 interface Page {
   url: string | false;
   fileSlug: string;
@@ -11,16 +12,19 @@ interface Page {
 /**
  * Removes leading id (e.g. `001-`) from blog filenames.
  */
-export const getBlogSlug = (page: Page) => page.fileSlug.replace(/^(\d*-)/, '');
+export function getBlogSlug(page: Page): string {
+  return page.fileSlug.replace(/^(\d*-)/, '');
+}
 
 /**
  * Indicates whether the given page is part of the JS API documentation.
  */
-export const isTypedoc = (page: Page) =>
-  page.url ? page.url.startsWith('/documentation/js-api/') : false;
+export function isTypedoc(page: Page): boolean {
+  return page.url ? page.url.startsWith('/documentation/js-api/') : false;
+}
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export default function pagesPlugin(eleventyConfig: any) {
+export default function pagesPlugin(eleventyConfig: any): void {
   // filters...
   eleventyConfig.addLiquidFilter('getBlogSlug', getBlogSlug);
   eleventyConfig.addLiquidFilter('isTypedoc', isTypedoc);
