@@ -5,6 +5,7 @@ import {typogrify} from 'typogr';
 
 import {markdownEngine} from './engines';
 
+/** A generator for filler text. */
 const lorem = new LoremIpsum({
   random: seedrandom("Feelin' Sassy!"),
 });
@@ -14,7 +15,7 @@ const lorem = new LoremIpsum({
  *
  * @see https://github.com/knicklabs/lorem-ipsum.js
  */
-export const getLorem = (type: string, number = 1) => {
+export function getLorem(type: string, number = 1): string {
   switch (type) {
     case 'sentence':
     case 'sentences':
@@ -27,7 +28,7 @@ export const getLorem = (type: string, number = 1) => {
       return lorem.generateWords(number);
   }
   return '';
-};
+}
 
 /**
  * Strips leading whitespace from each line of a string,
@@ -36,7 +37,7 @@ export const getLorem = (type: string, number = 1) => {
  * @see https://github.com/sindresorhus/strip-indent
  * @see https://github.com/jamiebuilds/min-indent
  */
-export const stripIndent = (contents: string) => {
+export function stripIndent(contents: string): string {
   // Find leading whitespace of first line (ignoring initial newlines)
   const match = /^[\n\r]*([ \t]*)(?=\S)/.exec(contents);
   if (match?.[1]?.length) {
@@ -47,27 +48,30 @@ export const stripIndent = (contents: string) => {
     );
   }
   return contents;
-};
+}
 
 /**
  * Truncates an HTML string without breaking tags.
  *
  * @see https://github.com/oe/truncate-html
  */
-export const truncateHTML = (html: string, words = 170) =>
-  truncate(html, words, {byWords: true, keepWhitespaces: true});
+export function truncateHTML(html: string, words = 170): string {
+  return truncate(html, words, {byWords: true, keepWhitespaces: true});
+}
 
 /**
  * Renders block of Markdown into HTML.
  */
-export const markdown = (content: string) =>
-  markdownEngine.render(stripIndent(content));
+export function markdown(content: string): string {
+  return markdownEngine.render(stripIndent(content));
+}
 
 /**
  * Renders single line of Markdown into HTML, without wrapping `<p>`.
  */
-export const markdownInline = (content: string) =>
-  markdownEngine.renderInline(content);
+export function markdownInline(content: string): string {
+  return markdownEngine.renderInline(content);
+}
 
 /**
  * Applies various transformations to plain text in order to yield
@@ -75,21 +79,26 @@ export const markdownInline = (content: string) =>
  *
  * @see https://github.com/ekalinin/typogr.js
  */
-export const typogr = (content: string) => typogrify(content);
+export function typogr(content: string): string {
+  return typogrify(content);
+}
 
 /**
  * Appends full page URL to internal links (for embedding in another page).
  */
-export const replaceInternalLinks = (content: string, url: string) =>
-  content.replace(/href="#/g, `href="${url}#`);
+export function replaceInternalLinks(content: string, url: string): string {
+  return content.replace(/href="#/g, `href="${url}#`);
+}
 
 /**
  * Checks if a given string starts with a comparison string.
  */
-export const startsWith = (str: string, check: string) => str.startsWith(check);
+export function startsWith(str: string, check: string): boolean {
+  return str.startsWith(check);
+}
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export default function typePlugin(eleventyConfig: any) {
+export default function typePlugin(eleventyConfig: any): void {
   // filters...
   eleventyConfig.addLiquidFilter('truncateHTML', truncateHTML);
   eleventyConfig.addLiquidFilter('markdown', markdown);
